Added configuration files and readme

+++ good/include/good/file.h
@@ -45,12 +45,12 @@ namespace good
         //--------------------------------------------------------------------------------------------------------
         /// Get file size. Returns FILE_OPERATION_FAILED if file doesn't exists.
         //--------------------------------------------------------------------------------------------------------
-        static int file_size( const TChar* szFileName );
+        static long file_size( const TChar* szFileName );
 
         //--------------------------------------------------------------------------------------------------------
         /// Read bytes from position iPos of the file in given buffer. Return false if file doesn't exists.
         //--------------------------------------------------------------------------------------------------------
-        static int file_to_memory( const TChar* szFileName, void* pBuffer, int iBufferSize, long iPos = 0 );
+        static long file_to_memory( const TChar* szFileName, void* pBuffer, long iBufferSize, long iPos = 0 );
 
         //--------------------------------------------------------------------------------------------------------
         /// Make folders for a file if they don't exist.
